refactor(SearchToolProvider): #408 lift 24 formatter files to ServicesModels — closes #408
The last remaining behavioural import in SearchToolProvider was
`Services` for the formatter family (Markdown, Text, JSON, HIG.*,
Frameworks.*, Unified.*, Footer.*, Sample.Format.*). All 24 formatter
files lift to ServicesModels in one mechanical move; they're pure
value types (structs conforming to Services.Formatter.Result) with
foundation-only deps.

After this PR, SearchToolProvider has zero behavioural cross-package
imports. #408 closes.

## File moves

24 files from Sources/Services/Formatters/ → Sources/ServicesModels/Formatters/:

- Services.Formatter.Result.swift (the protocol)
- Services.Formatter.StringExtensions.swift (helpers)
- Services.Formatter.{JSON, Markdown, Text}.swift (main search renderers)
- Services.Formatter.HIG.{JSON, Markdown, Text}.swift
- Services.Formatter.Frameworks.{JSON, Markdown, Text}.swift
- Services.Formatter.Unified.{JSON, Markdown, Text}.swift
- Services.Formatter.Footer.{Formattable, Item, Kind, Markdown, Provider, Search, Text}.swift
- Sample.Format.{JSON, Markdown, Text}.Search.swift

Stale `import ServicesModels` (self-import) and `import SampleIndex`
(replaced by SampleIndexModels) trimmed from each file. Final shared
import set: Foundation, SampleIndexModels, SearchModels,
SharedConstants, SharedCore.

## SearchToolProvider

- Drops `import Services` from Sources/SearchToolProvider/CompositeToolProvider.swift
- Drops the 2-argument convenience init (which constructed concrete
  service actors). It moves to the test side as
  `CompositeToolProvider+ServicesWiring.swift` in two copies — one
  in SearchToolProviderTests, one in ServeTests — so the 30 existing
  test callsites keep their two-argument shape. The two test-target
  copies are identical; Swift test targets can't share extension
  files across directories without play.
- Final production import set (foundation-only): Foundation, MCPCore,
  MCPSharedTools, SampleIndexModels, SearchModels, ServicesModels,
  SharedConstants, SharedCore, SharedUtils.

## Package.swift

- SearchToolProvider target drops `Services` dep
- SearchToolProviderTests gains SampleIndex + Services deps for the
  test-side wiring helper
- ServeTests gains SampleIndex + SampleIndexModels + Services +
  ServicesModels + SharedConstants deps for the parallel wiring helper
- ServeTests's 3 callsites with the now-removed 2-arg init keep
  working via the local helper file

## Acceptance grep

$ grep -rln '^import Services$\|^import SampleIndex$\|^import Search$' Packages/Sources/SearchToolProvider/
(empty)

## Verification

xcrun swift build: clean
xcrun swift test: 1441 tests in 161 suites pass

## #408 closes

SearchToolProvider's final import set is foundation-layer + Models
targets only. The target compiles and tests in isolation; the CLI
composition root constructs concrete service actors and passes them
across protocol seams (DocsSearcher, Sample.Search.Searcher,
Teaser, UnifiedSearcher).

DI epic #381: 2 of the 3 heavy-tier children now closed (#400
Search, #408 SearchToolProvider). #403 Indexer remains as the
final architectural slice.
