+++ Packages/Sources/ServicesModels/Formatters/Sample.Format.JSON.Search.swift
@@ -1,6 +1,4 @@
-import ServicesModels
 import Foundation
-import SampleIndex
 import SampleIndexModels
 import SharedConstants
 import SharedCore
