+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.Footer.Markdown.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 
 // MARK: - Footer Markdown Formatter
