+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.Footer.Search.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 import SharedConstants
