+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.Footer.Text.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 
 // MARK: - Footer Text Formatter
