+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.HIG.JSON.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 import SearchModels
 import SharedCore
