+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.HIG.Markdown.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 import SearchModels
 import SharedConstants
