+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.Markdown.swift
@@ -1,4 +1,3 @@
-import ServicesModels
 import Foundation
 import SearchModels
 import SharedConstants
