+++ Packages/Sources/ServicesModels/Formatters/Services.Formatter.Unified.JSON.swift
@@ -1,6 +1,4 @@
-import ServicesModels
 import Foundation
-import SampleIndex
 import SampleIndexModels
 import SearchModels
 import SharedConstants
